+++ source/Tests/QA/ResourceSchema.common.Tests.ps1
@@ -7,9 +7,23 @@ param (
     $SourceManifest
 )
 
-
-$Principal = [Security.Principal.WindowsPrincipal]::new([Security.Principal.WindowsIdentity]::GetCurrent())
-$isAdmin = $Principal.IsInRole([Security.Principal.WindowsBuiltInRole]::Administrator)
+if ($isLinux -or $IsMacOS)
+{
+    $skipTests = $true
+}
+else
+{
+    $Principal = [Security.Principal.WindowsPrincipal]::new([Security.Principal.WindowsIdentity]::GetCurrent())
+    $isAdmin = $Principal.IsInRole([Security.Principal.WindowsBuiltInRole]::Administrator)
+    if ($isAdmin)
+    {
+        $skipTests = $false
+    }
+    else
+    {
+        $skipTests = $true
+    }
+}
 Describe 'Common Tests - Script Resource Schema Validation' -Tag WindowsOnly {
     if ($isAdmin -and ($IsWindows -or $PSVersionTable.PSEdition -eq 'Desktop'))
     {
@@ -23,11 +37,11 @@ Describe 'Common Tests - Script Resource Schema Validation' -Tag WindowsOnly {
         Context $scriptResourceName {
             $scriptResourcePath = Join-Path -Path $dscResourcesFolderFilePath -ChildPath $scriptResourceName
 
-            It 'Should pass Test-xDscResource' -Skip:(!$isAdmin -or !($IsWindows -or $PSVersionTable.PSEdition -eq 'Desktop')) {
+            It 'Should pass Test-xDscResource' -Skip:$skipTests {
                 Test-xDscResource -Name $scriptResourcePath | Should -Be $true
             }
 
-            It 'Should pass Test-xDscSchema' -Skip:(!$isAdmin -or !($IsWindows -or $PSVersionTable.PSEdition -eq 'Desktop')) {
+            It 'Should pass Test-xDscSchema' -Skip:$skipTests {
                 $mofSchemaFilePath = Join-Path -Path $scriptResourcePath -ChildPath "$scriptResourceName.schema.mof"
                 Test-xDscSchema -Path $mofSchemaFilePath | Should -Be $true
             }