+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Now skipping isAdmin on non-windows (because creds works differently)
+
 ## [0.4.0] - 2019-12-04
 
 ### Added